Update adHocSubmit
This modification adds a call to e.range.clearContent() at the end of each conditional block, which clears the value of the edited cell (either the "Submit" or "Clear" cell, depending on the block that was executed).

The code is triggered by the onEdit() function, which is only called when a user edits a cell in a sheet. If you're trying to run this script as a standalone script, you'll need to modify it to include a function that can be triggered manually (e.g. function submitData() { ... }).

The script assumes that the user will only click the "Submit" or "Clear" buttons on the "submit" sheet. If the user interacts with the sheet in a different way (e.g. edits a cell without clicking a button), the script will not be triggered.

There are some magic numbers in the code (e.g. 24, 4) that may not be appropriate for all use cases. Consider defining constants or variables for these values to make the code more readable and flexible.